The power consumption of time extended by nvram is regarded as cpu idle
Previous implementation caculated the power consumption based on active
cpu power for the time duration extended by slow memory, NVRAM. Now,
the extended time is caculated based on cpu idle power.
A new HM checked policy is added but not merged into simrts binary.
That policy is almost same with HM or DRAM depending on the scaling
factor of memory.